// src/interfaces/master-server-list.ts
export interface MasterServerListEntry {
  addr?: unknown;
  steamid?: unknown;
  version?: unknown;
  name?: unknown;
  ip?: unknown;
  port?: unknown;
  region?: unknown;
  gm?: unknown;
  mods?: unknown;
  password?: unknown;
  whitelist?: unknown;
}

export interface MasterServerListResponse {
  successful?: unknown;
  data?: unknown;
}